fix: cumulative sum transform

+++ insights/insights/doctype/insights_query/utils.py
@@ -203,7 +203,9 @@ def apply_cumulative_sum(columns, results):
     results_df = pd.DataFrame(results[1:], columns=column_names)
 
     for column in columns:
-        results_df[column.get("label")] = results_df[column.get("label")].cumsum().fillna(0)
+        results_df[column.get("label")] = (
+            results_df[column.get("label")].astype(float).cumsum().fillna(0)
+        )
 
     return [results[0]] + results_df.values.tolist()
